requirements.txt
python-telegram-bot==20.7
Flask==3.0.3
pandas==2.2.2
python-dateutil==2.9.0.post0
Unidecode==1.3.8
requests==2.32.3
